+++ src/test/java/dagger/sheath/InjectMock.java
@@ -1,4 +1,4 @@
-package dagger.sheath.junit;
+package dagger.sheath;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
