+++ src/test/java/dagger/sheath/NopSupplier.java
@@ -1,4 +1,4 @@
-package dagger.sheath.junit;
+package dagger.sheath;
 
 import java.util.function.Supplier;
