+++ src/test/java/dagger/sheath/ProviderWrapper.java
@@ -1,4 +1,4 @@
-package dagger.sheath.junit;
+package dagger.sheath;
 
 import dagger.internal.DoubleCheck;
 import lombok.extern.slf4j.Slf4j;
