+++ src/test/java/dagger/sheath/junit/DaggerExtension.java
@@ -1,5 +1,9 @@
 package dagger.sheath.junit;
 
+import dagger.sheath.CtxWrapper;
+import dagger.sheath.EventHandler;
+import dagger.sheath.InjectMock;
+import dagger.sheath.NopSupplier;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.reflect.FieldUtils;
