--- src/test/java/dagger/sheath/junit/Fields.java
@@ -1,4 +1,0 @@
-package dagger.sheath.junit;
-
-public class Fields {
-}
